// src/app/models/EmployeeLog.ts
export class EmployeeLog {
    employeeId?: string;
    employeeNumber?: string;
    employeeNameFl?: string;
    employeeNameSl?: string;
    civilId?: string;
    logDate?: Date;
    logTime?: string;
    logTypeId?: string;
    logTypeNameFl?: string;
    logTypeNameSl?: string;
    id?: string;
}